// src/app/interfaces/conductor.ts
export interface Conductor {
    id_conductor?: number;
    id_dispositivo?: number;
    nombre?: string;
    apellido?: string;
    nombre_vehiculo?: string;
    pais?: string;
    estado?: string;
    direccion?: string;
    tipo_sangre?: string;
    telefono?: string;
    correo?: string;
    contacto_emergencia?: string;
    licencia?: string;
    tipo_licencia?: string;
    id_usuario?: number;
}